uetpark/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
Pillow==10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7 